Database to binary. File permissions. Small fixes.
Files are now in binary form, not base64.

Set exec perms on files before packaging.

Config was world readable. While it's not too big of an issue,
it could be a security issue.

Fixed the pid finding of plugins, but still isnt good enough.

Fixed init script for finding full paths (init should be redone)